requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# LangChain + LangGraph
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
langgraph==0.1.5

# BM25
rank-bm25==0.2.2

# Redis
redis[asyncio]==5.0.4

# LLMLingua (context compression)
llmlingua==0.2.2

# LangSmith monitoring
langsmith==0.1.77

# OpenTelemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0

# HTTP client for MCP tools
httpx==0.27.0